start optical flow

+++ drivers/nvic.c
@@ -1,6 +1,7 @@
 /* Includes ------------------------------------------------------------------*/
 #include "nvic.h"
 #include "conf.h"
+#include "tools.h"
 
 uint16_t tim1_count=0;
 uint16_t tim2_count=0;
@@ -152,8 +153,8 @@ void TIM2_IRQHandler(void)
 
 void TIM3_IRQHandler(void)
 {
-  State_Update((float)1/TIM3_Freq);             // 状态信息更新 Pitch, Roll, Yaw , etc.
-  PID_Update((float)1/TIM3_Freq);        // PID控制输入每次循环时间
+  State_Update(1.0f/TIM3_Freq);             // 状态信息更新 Pitch, Roll, Yaw , etc.
+  PID_Update(1.0f/TIM3_Freq);        // PID控制输入每次循环时间
   MORTOR_Output();
   if(TIM_GetITStatus(TIM3, TIM_IT_Update) != RESET) {
     TIM_ClearITPendingBit(TIM3,TIM_IT_Update);
